added loop code to constructor
now it will remain "on" when it doesn't have a rom loaded. Also moved the loadROM() function to cpu class as streamLoadFromFile()

+++ refactored_chip_8/refactored_chip_8/bus.h
@@ -40,6 +40,6 @@ public:
 
 	theme& currentTheme;
 
-	int lastKeyPress;
-	int currentHeldKey;
+	uint8_t lastKeyPress;
+	uint8_t currentHeldKey;
 };
